Makes http transport handle single and batches messages differently
Single messages don't need to provide any message length headers since it can assume that the whole payload is the message. By splitting the logic for the two it is able to process single messages with less overhead, and it also makes the server potentially capable of processing messages from an alternative client implementation that does not support batching.